Avoid a race condition with menus being rewritten on app startup
A few applications, such as Sublime Text, seem to rewrite their menus
on startup after loading plugins. This can race with Afloat which
assumes that the menus stay static. As a stopgap measure, wait a bit
before inserting Afloat's items.

It would be better to watch for menu changes and reinstall our items
then, as this only papers over the issue at startup. Later changes
to the menus can (and do) trash Afloat's additions.